Update
Changing the backend a bit. Ran into naming issues using 'issue' for a name so renaming to 'bug' so theyre bug reports now not Issues. Should have been that way from the start really. Also simplifying the data a bit by removing the complexity I created when I had comments and bugs user the same field. It was originally intended for a flatdb so I tried to make it simple and it ended up complex.

Havent created controllers or routes yet